+++ src/Bridges/Constraint/interval.jl
@@ -86,14 +86,14 @@ end
 function MOI.get(
     ::SplitIntervalBridge{T,F,S,LS},
     ::MOI.NumberOfConstraints{F,LS},
-) where {T,F,S,LS}
+)::Int64 where {T,F,S,LS}
     return 1
 end
 
 function MOI.get(
     ::SplitIntervalBridge{T,F,S,LS,US},
     ::MOI.NumberOfConstraints{F,US},
-) where {T,F,S,LS,US}
+)::Int64 where {T,F,S,LS,US}
     return 1
 end
