+++ src/Bridges/Constraint/quad_to_soc.jl
@@ -189,7 +189,7 @@ function MOI.get(
         MOI.VectorAffineFunction{T},
         MOI.RotatedSecondOrderCone,
     },
-) where {T}
+)::Int64 where {T}
     return 1
 end
