+++ src/Bridges/Constraint/scalarize.jl
@@ -58,7 +58,7 @@ end
 function MOI.get(
     bridge::ScalarizeBridge{T,F,S},
     ::MOI.NumberOfConstraints{F,S},
-) where {T,F,S}
+)::Int64 where {T,F,S}
     return length(bridge.scalar_constraints)
 end
