+++ src/Bridges/Constraint/set_map.jl
@@ -64,7 +64,7 @@ end
 function MOI.get(
     ::SetMapBridge{T,S2,S1,F},
     ::MOI.NumberOfConstraints{F,S2},
-) where {T,S2,S1,F}
+)::Int64 where {T,S2,S1,F}
     return 1
 end
