+++ src/Bridges/Constraint/square.jl
@@ -167,14 +167,14 @@ end
 function MOI.get(
     ::SquareBridge{T,F,G,TT},
     ::MOI.NumberOfConstraints{F,TT},
-) where {T,F,G,TT}
+)::Int64 where {T,F,G,TT}
     return 1
 end
 
 function MOI.get(
     bridge::SquareBridge{T,F,G},
     ::MOI.NumberOfConstraints{G,MOI.EqualTo{T}},
-) where {T,F,G}
+)::Int64 where {T,F,G}
     return length(bridge.sym)
 end
